Add command for running tests with -X jit-auto
Summary:
As I'm working on AutoJIT + type profiling, I'm finding that there are other
test failures that relate purely to AutoJIT, so adding that as a supported way
of running tests.

Most of the test failures I was seeing were JIT tests that expected functions to
be compiled after being called once, this is not true with AutoJIT.  This diff
adds a `cinderjit` function to read what the current AutoJIT threshold is so we
can tell if we're running with it or not.  It'll be zero when AutoJIT is not
enabled.

There is still an outstanding failure with a StaticPython test that I've tasked
up in T165854755.

Reviewed By: mpage

Differential Revision: D50021611

fbshipit-source-id: 0e2d8a15112251e0969b34859fccf87e85afd06b