Optimize generator creation
Summary:
In `InitialYield`'s codegen, we were framing the `rep movsq` with
`std` and `cld`, to have the copy progress from higher to lower addresses. If
we start the pointers at the bottom of their respective buffers rather than the
top, we can leave `DF` alone and have the copy go from lower to higher
addresses (the System V ABI requires that DF be cleared on function entry/exit,
so it's safe to assume it's 0 coming into this code).

Reviewed By: jbower-fb

Differential Revision: D49653637

fbshipit-source-id: ea68e3ed83536fcb996d9478179cf82457e6818a