introduce cinderhooks, start with type create/dealloc
Summary:
Cinder features rely on a number of hooks we've patched into core CPython. We upstreamed the most important of these into 3.12, but we didn't get everything; some new hooks we actually added started using after we'd created our upstream-hooks plan, and some we just missed in the initial pass.

This diff adds a simple-as-possible strategy for providing these hooks for now: just a set of global function pointers, which CinderX can set to the appropriate CinderX functions. We use the `Ci_` prefix for these symbols, and collect them in a new `cinderhooks` module, so we can easily see what hooks we need to upstream, or port to a new Python version.

Here we just add the new module, and a pair of hooks, for type creation and dealloc.

We use `PyUnstable_GC_VisitObjects` to register any types created before `Cinder_Init`, similar to how we do for functions.

The question might be asked, why are we adding this to CPython upstream directories (`Include/` and `Python/`) rather than to a Cinder-specific directory (e.g. `Cinder/`)? The reason is that the only other compilation unit in `Cinder/` (`Cinder/cinder.cpp`) is something that belongs in CinderX eventually; this is parallel to e.g. `Jit/`, `Shadowcode/`, `StaticPython/`, `StrictModules/`, etc; these are all currently optionally either compiled into CinderX extension, and/or into Python itself, but should eventually be part of only CinderX. But the compilation unit I'm adding here should specifically only ever be part of core Python, it can't/shouldn't ever be compiled into CinderX. It's simply a patch we need to carry against upstream, for now. So I think it would be confusing to add it to `Cinder/`. We could add a new `CinderNoCinderX/` top-level directory, but tbh just putting it in `Python/`, with the name `cinder` in the filename, seems equally clear.

Reviewed By: mpage, swtaarrs

Differential Revision: D49343986

fbshipit-source-id: 9c559bf1df083ef5563b8b7a16d91130fa1a3cbb