Update scripts to allow Python tests to move to CinderX
Summary:
This enables us to move Python tests out of `Lib/test/` and into `CinderX/test_cinderx/`. I've left backward compatible make rules so most old flows like `make -j testcinder` will still work, but tests previously run individually like this:
```
$ ./python -m unittest test.test_shadowcode
or
$ ./python -m test test.test_shadowcode
```
Now need to respectively be run as:
```
$ cd CinderX && ../python -m unittest test_cinderx.test_shadowcode
or
$ cd CinderX && ../python -m test --testdir $(pwd) test_cinderx.test_shadowcode
```

The primary changes are:
* Lifting the Cinder Python test parts of the global `Makefile[.pre.in]` and into a new `CinderX/Makefile`.
  * The goal is primarily making it possible to hack on the CinderX test running code without having to change the CPython code.
  * A secondary goal is to minimize dependencies on CPython build scripts although there are still some minor ones (e.g. something needs to set build a Python binary and set it in `TESTPYTHON`).
* Making it possible to use the `jit_test_runner.py` script (now `cinder_test_runner.py`) to run all CinderX tests. Before we used the default `run_test.py` script for `testcinder` and `testcinder_refleak`. Again this minimizes future needs to change CPython scripts to make tests run.
* The logic to decide which tests to skip under certain conditions (e.g. under ASAN, w/wo JIT, etc.) are now in Python in `cinder_test_runner.py` with fewer hardcoded in that script. Overall this should be much cleaner/more unified now.

In this diff we only move `test_shadowcode` to demonstrate correctness. A follow-up moves the rest.

Something NOT addressed in this diff is making it possible to test Cinder features which are in our CPython fork without the scripts now in CinderX. It's not much work to do this, but I'm not sure it buys much. The best implementation is also unclear until we switch CI systems with the move to third-party/fbcode.

Reviewed By: itamaro

Differential Revision: D50354346

fbshipit-source-id: 694f6716598c6bcc41202d46dfa13b9d23908278