Move C++ Strict Modules/Runtime tests into CinderX/
Summary:
Mostly this diff is renaming files but it also lifts the build/run logic out of the core `Makefile[.pre.in]` into `CinderX/Makefile`. As with moving the Python tests, the idea is to minimize the need to change CPython code when changing how these tests are run.

The dependencies between the core build and the new Makefile is pretty ugly, but at some point we'll probably replace this with another build system (likely Buck). So I think making it cleaner is not worth the effort at this point.

Reviewed By: alexmalyshev

Differential Revision: D50516544

fbshipit-source-id: 02cbfeb23426b5c7c92de7b3a671abcff7e60556