Delete cinderjit profile tests that rely on classmethod descriptor chaining
Summary:
upstream issue [GH-89519](https://github.com/python/cpython/issues/89519) is about removing classmethod descriptor chaining capability, which was added in 3.9, deprecated in 3.11, and pending removal in 3.13.

we have code that breaks due to the "new" (to us, when upgrading from 3.8 ) descriptor chaining behavior, so instead of figuring out how to fix it, considering this is deprecated and being removed in upstream, it seems preferable to "accelerate" that removal internally by backporting the deprecation and removal in cinder 3.10.

the stacked two diffs backport the deprecation (D50534764) and removal (D50534886). this diff deletes a couple of tests in cinder that rely on that behavior.

Reviewed By: swtaarrs

Differential Revision: D50578116

fbshipit-source-id: 2faed200a3a130ff7be7d8c6a231c6431789b330