Simplify CondBranch conditions that come from a widened IntConvert
Summary:
CondBranch sets the zero flag based on its given unboxed integer.  It doesn't
care about what size the integer is, and if this integer was widened from
another integer doesn't change the fact of it being zero or non-zero.

Currently we often generate:

  v2 = IsTruthy v1:Bool
  CondBranch<1, 2> v2

and then simplify the IsTruthy down to:

  v3 = LoadConst<Py_True>
  v4 = PrimitiveCompare<Equal> v1, v3
  v5 = IntConvert<TCInt32> v4

which is where we get this pattern from.

Reviewed By: swtaarrs

Differential Revision: D50558483

fbshipit-source-id: 08e3bc1def8058772edd6d4e524d788e7c17109d