Improve handling and logging of failed type loads in preloader
Summary:
We have more reports of some crashes during startup with the preloader.  This is likely due to static pyc invalidation issues as the stack trace points at resolving the types for checking arguments.

This fixes up that case and a couple of others so that we'll check that the types resolve.  It also starts logging out the name of the function that we're trying to compile which should point us at what pyc needs to be invalidated.

The end result is that we should just skip pre-compilation, and then we'll probably report the error later if these methods actually get invoked.

Reviewed By: Kronuz

Differential Revision: D50744277

fbshipit-source-id: 3ae8c364f8ceb135231c3216a2e2878a09b86a21