Spawn threads on-demand for parallel marking
Summary:
Thread pools and forking do not play well together. In order to
keep the previous thread pool based approach we would need to respawn
the thread pool in each child process after fork. This is possible
but feels error prone and easy to get wrong. Instead, we can spawn
the worker threads on-demand when performing marking. This seems
to be roughly the same performance-wise, and leads to a slightly
simpler implementation.

Reviewed By: alexmalyshev

Differential Revision: D49065836

fbshipit-source-id: 4f108db8c09d3b51b5db97ccee083c2a075a7797