Move parallel GC into Cinder
Summary:
Move the implementation of the parallel GC out of core CPython and
into Cinder to minimize coupling. To achieve this we introduce two
small hooks into core CPython for setting a custom GC.

Although there are a lot of lines changed there is very little new
functionality. The vast majority of the code has already been reviewed
earlier in the stack. Please focus on the following:

- New hooks added to `cinder/exports.h` and changes to gcmodule.c to
  use them.
- Parallel GC APIs added to `cinder/cinder.h` and their implementations
  in parallel_gc.cpp.
- Small change in `cinder.cpp` to appease the linker.

Reviewed By: jbower-fb

Differential Revision: D50659008

fbshipit-source-id: d22419c5bd6e96e9acd7bd5a21a55283b718a077