Wait until all worker threads have exited before finalizing parallel gc state
Summary:
During finalization, the interpreter will perform a final collection
immediately before destroying GC state. Depending on the vagaries of the
OS scheduler, we may reach this point before some worker threads have
been woken up. When that occurs, they will still need access to the
synchronization primitives in the barrier that determines when a collection
is done. We must wait for them to proceed past this point before we can
destroy the parallel gc state that they share.

Reviewed By: DinoV

Differential Revision: D50978313

fbshipit-source-id: f05c1d358759bad15219434228f82e4f1f41241b