Ensure that GCd functions get removed from trampoline list
Summary:
Similar to the actual JIT, we build up a list of functions for which we generate a trampoline pre-fork.

If a registered function gets GCd before generating a trampoline, we end up seg-faulting.

This diff uses the existing function watcher to remove such entries from the list of registered functions.

Reviewed By: carljm

Differential Revision: D51055358

fbshipit-source-id: 05581cabda6d5eaecd7706c4777ae1bc840d3c6f