Support calling without "self" on patched INVOKE_FUNCTIONs
Summary:
Currently when we can resolve a function that is statically invoked to a specific type (because the type is final, the function is final, or we have an exact type) we generate an `INVOKE_FUNCTION` and explicitly pass `self` as the first parameter.

When the function is patched with a mock in place this results in an extra parameter being passed vs the non-static Python case.

This adds a knob to cinder which causes us to peel off the first argument and bind to a descriptor if the patch is a descriptor, or drop the first argument entirely if it's not.

We can then enable the knob and fix all of the tests in IG which are expecting the current behavior.

Reviewed By: carljm

Differential Revision: D51212209

fbshipit-source-id: a097adb1a32c0322f7c84b2b8608786336329356