Put some effort into CinderX shutdown
Summary:
This fixes some issues with tests which re-inititalize Python and cause issues when we initialize CinderX later in the following changes.

This is still very much a best effort approach.

The biggest change is freeing Shadow Code resources. I didn't want to do this originally as Shadow Code is a dead-end, but not doing this caused ASAN issues with some runtime tests which re-start Python. Fortunately much of this code already existed in Python as part of the refleaks testing infrastructure, and I've just converted it to C.

Reviewed By: oclbdk

Differential Revision: D50866604

fbshipit-source-id: 8646c375a68df85e4bd7aac0a9d51ed61db9de81