Change imports from _static to cinderx.static
Summary: This pairs up with a later change where we perform CinderX initialization on `cinderx` module load. Changing the imports like this means anything using Static Python will implicitly ensure CinderX is initialized. This is primarily to help a few tests which suppress using "site" start-up scripts which we use to make sure CinderX is loaded for most applications.

Reviewed By: oclbdk

Differential Revision: D50960604

fbshipit-source-id: 639274501933368531668def55c5ad7f6cfc412e