Tidy-up build scripts
Summary:
The primary motivation behind this is I think it's become confusing where to put things in a CInder vs. CinderX world c.f. the `Cinder/` directory which this series of changes removes. A secondary concern is the Makefile header dependencies are almost certainly wrong anyway because this is such a confusing mess. I don't guarantee they're correct now but it should be better.

The main thing this diff is doing is grouping everything "CinderX" into one lump rather than separating out Jit, Static Python etc. This simplifies things considerably. The other thing it's doing is laying the ground work for moving things like `Jit/` etc. directly into `CinderX/` but explicitly specifying the paths as being relative to `CinderX/`.

Reviewed By: alexmalyshev

Differential Revision: D51103492

fbshipit-source-id: 8eee327037a7ca3dda240452ba5e1bdc28aecb21