Remove Cinder/
Summary:
The `Cinder/` directory is just confusing. My original intent was it would be a place to put "Cinder" code to help separate it out from CPython. However, that never happened and it ended up (before this set of changes) containing a mix of stuff that needs to be in CPython and others that will be part of CinderX. Earlier changes removes chunks of the content, and this gets rid of the rest and removes the directory entirely.

Most of the remaining content was related to initialization and is moved to `_cinderx.cpp` (the native module code for CinderX). The other piece was `cinder.h` which contained no-longer needed prototypes for initialization functions, and the API for the ParallelGC which should be in its own file.

Reviewed By: carljm

Differential Revision: D51103507

fbshipit-source-id: abd1f376589cb07ffe726e01318797debeb126b8