Prepare to remove the CinderX/[Jit, Shadowcode, StaticPython, StrictModules, ThirdParty] symlinks
Summary:
Clears the way to move the major C++ code directories from the root to `CInderX/`, getting rid of the symlinks. This diff cleans up the few places which were relying on the root-based location. The actual move happens in the next diff.

Note not all symlinks are gone yet and so there are still things like a root-based include paths in things like the `cinder.c` module in `Modules/`. These will need to be cleaned up later.

Reviewed By: carljm

Differential Revision: D51146320

fbshipit-source-id: 5787dc55419fd93a3caaacea587e6f2bcaf7c9cb