extract CheckedList to CinderX
Summary:
Extract `CheckedList` type and everything it needs out of `listobject.c` and into a new `StaticPython/checked_list.c`.

"Everything it needs" includes most of the static functions in `listobject.c`, so rather than adding upstream `Ci_*` exports of all of these, `StaticPython/checked_list.c` now includes a copy of much of `listobject.c`.

I chose to minimize changes to the copied code in this diff; I only made the minimal changes necessary for it to compile and run (adjusting a few type checks and the names of `PyCheckedListIter_Type` and `PyCheckedListRevIter_Type`.)

We could make much more extensive changes in the service of clarity and easier debugging; e.g. rename everything to `chklist_*` instead of `list_*`. The tradeoff here would be more difficult porting of future upstream list changes.

We could just only port future upstream list changes if we really want them, and accept that otherwise `CheckedList` will remain a frozen-in-time copy of Python 3.10 list. The exception would be any upstream changes that impact the layout of the list object itself; these we would have to port. (Such changes seem unlikely at this point, though; list is quite stable.)

Reviewed By: oclbdk, DinoV

Differential Revision: D51261559

fbshipit-source-id: ef4dc1fb1f7ccf6595604393738a8653d41d5cb4