remove SEQUENCE_REPEAT
Summary:
This is a highly specialized opcode that was added to enable regression-free conversion of one specific highly-tuned Cython function, which is now gone anyway. This isn't typically a common operation in hot code.

If we ever want to reintroduce this optimization in Static Python, it would be better handled as an invoke of a type slot method, rather than as a dedicated opcode.

Reviewed By: DinoV

Differential Revision: D51275210

fbshipit-source-id: a7fe2f4d19399bbdbb7b76154f8cf89697751259