add PyModule_Dict as a hook
Summary:
Adds `PyModule_Dict` as a hook. Discussed with Carl; looked into the option of removing this hook altogether, but there are a few use cases in `import.c` that still rely on branching at this function: https://fburl.com/code/c6eamlmq (*attempted this here: D51317520*).

There might be a better path to upstreamability, but wanted to throw this out for feedback.

Reviewed By: carljm

Differential Revision: D51320293

fbshipit-source-id: 4fe651e02a5c77967867e977d9d97a771687c686