Fix shadowcode finalization
Summary:
Below is copied from D51711697, this is a sync of that change to the Cinder repo.

Based on the stacktrace in P893665304, there's a null pointer deference lurking
here.  The trace suggests that we either have a null code object, or a code
object with a null co_consts field.  I can't tell which so I'm adding checks for
both and it fixes the crash.

Reviewed By: jbower-fb

Differential Revision: D51737049

fbshipit-source-id: cdc2c487089825af6fc48050db91833ed2feb78d