Work around TSAN-flagged issue when reading from a shared dict during multi-threaded compilation
Summary:
`PyDict_GetItem()` saves and restores the current exception state as part of ignoring errors which might occur during dictionary look-ups. These should be unlikely in our case as we're during a lookup on a dictionary where the keys should all be regular Python strings (so no calls to a Python hashing function, or Python comparison function during probes).

This situation is unlikely but I think not impossible? Should we do something better here?

Reviewed By: alexmalyshev

Differential Revision: D51918727

fbshipit-source-id: 19973a5a503b2e03df56ae5a8ffdc4ca5c7ddd0a