Properly initialize the JIT for runtime tests
Summary:
The way that the runtime tests initialize the JIT is not quite right.  It's
supposed to be done via `Py_Initialize()`, but currently the only way to tell
that to turn on the JIT is via CLI arguments or an environment variable, both of
which are hard to set.  There's also parts of RuntimeTests/main.cpp that
manually alloc and dealloc the global code allocator just to get things moving.

This diff adds a new field to the JIT config object that tells the JIT to force
initializing.  This is now set for all runtime tests by default, but it's
disabled for things like CmdLineTest where it's explicitly trying to test the
CLI and environment value logic.

Reviewed By: mpage

Differential Revision: D51954850

fbshipit-source-id: 87bd74b5da3af7e30abdc6644ff94962fba69c48