Remove function object from Preloader, map code objects to Preloaders
Summary:
The JIT code is inconsistent with how Preloader is initialized. Sometimes it's
initialized with a function object and sometimes it isn't.  Ultimately the
function object isn't used by the Preloader itself, it's just a convenient place
where we stash it and then later use it for the purposes of the multithreaded
compile codepath.

Keeping the function object on the Preloader is dangerous because while a code
object will stay the same, its function object can change over time (e.g. nested
functions).  We want the compiler to only make use of information on the code
object and not on the function object.

The multithreaded compile codepath keeps a map of "compilation unit" to
Preloader object, that gets read by the ThreadedCompileContext.  A compilation
unit is either a function or a code object.  To simplify the map and how it is
used, this diff changes its key to always be the code object, to enforce some
level of consistency. The function that access it support either type.

Reviewed By: carljm

Differential Revision: D51645729

fbshipit-source-id: f1faa5aab4de0434ee9907591d65bdcd73628047