eliminate special case for strict modules in _PyObject_LookupAttr
Summary:
This special case was added so that we'd suppress `ImportCycleError` in attribute lookups, which is necessary to handle some edge cases in lazy imports. But the fast-paths here should not implement different semantics from the main path, so the correct approach is to also suppress `ImportCycleError` in the slow path, too.

This means that `ImportCycleError` is no longer suppressed only on lookups against module objects, but on lookups against C extension types generally. I think this is not ideal, but OK. It means that in the rare case of a `hasattr/_PyObject_LookupAttr` lookup on a C extension type, where the attribute lookup has dynamic behavior that happens to trigger a lazy import and raise `ImportCycleError`, it will be suppressed and the `hasattr/_PyObject_LookupAttr` will just report "attribute not present." This parallels the existing behavior of `hasattr/_PyObject_LookupAttr` when a dynamic attribute calculation raises `AttributeError` internally.

This still won't happen on lookups against normal Python objects, because of the fast path for `PyObject_GenericGetAttr` where we don't currently suppress `ImportCycleError`. Maybe for consistency we should, but I think consistency isn't that important here, and given the choice it's probably better not to suppress.

Removing this special case means `Ci_strictmodule_lookupattro` is now unused, so I removed it.

I also happened to notice that the `suppress` argument of `Ci_module_lookupattro` is unused (and set to `1` at all callsites anyway), so I removed it.

Reviewed By: jbower-fb, Kronuz

Differential Revision: D52059168

fbshipit-source-id: 4fa5045a228b8cde7bff95a2f77ff49710ddf5ac