Refactor cinder module into cinder.py + _cinder.so + _cinderx.so
Summary:
Moves features of the `cinder` module which are primarily CinderX related over to the `_cinderx` module. To simplify migration, the native `cinder` module has been renamed to `_cinder` and a new Python version of `cinder` introduced which includes everything from `_cinder` and everything which used to be in the original module from `_cinderx`. Thus end-users won't see a change.

Something which is a bit of a gray area were functions which relied on being able to query Shadow Frames. On the one hand Shadow Frames are about optimizing JIT performance and part of CinderX. However, they are also queryable regardless of JIT. So there are functions in `cinder` which should work without CinderX being loaded even though they refer to Shadow Frames. I've chosen to leave these in the `cinder` module and refactor the shadow-frame API so more of it lives in the `pycore_shadow_frame.h` header in core Python. To support this I added a few more hooks to support the JIT side of shadow frames.

Reviewed By: carljm

Differential Revision: D52302194

fbshipit-source-id: b37487ebb5f186810afb4a270eba8dd041078b0f