Make sure _Py_ShadowFrame_WalkAndPopulate is in libpython
Summary:
In upcoming changes we'll be breaking many Cinder features out into a separate Python module. This would break users of this function as they currently expect to find it in libpython.

Although this symbol will now always be in libpython, it may not work as it relies on CinderX being loaded. In practice I think CinderX will always be loaded for the use cases I'm aware of. If CinderX is not present it returns valid empty data which should at least avoid a hard crash.

Reviewed By: carljm

Differential Revision: D52302640

fbshipit-source-id: ce99ef4f3a647fe82e6586d220b4578a2ae7a842