Fix fbsource import of cpython 3.12
Summary:
For complicated reasons the Meta source code syncing infrastructure cannot support .gitattributes macros, as documented here:

https://git-scm.com/docs/gitattributes#_defining_macro_attributes

Note especially this part:

> not in .gitattributes files in working tree subdirectories

The logic we use to extract a slice of the Meta monorepo and create a git repo for use on GitHub mangles the repo contents since this file is not at the "root" of the temporary exported git repo. Reworking our logic to fix this inside the code syncing logic turned out to not be possible without extended effort that is not worth it right now.

So instead we just backout these commits:

1. https://github.com/python/cpython/commit/0b906ae562fa057162a67d4a5b7ea426f6b9dd23
2. https://github.com/python/cpython/commit/f1e559b7544d665a84747a7bc7ecebadef2a6be2

(I didn't ACTUALLY revert them, just fixed the lines in-place (see the rearranging done in the second commit))

Reviewed By: itamaro

Differential Revision: D52377523

fbshipit-source-id: 0307ab4bbfbcd45ffeeb1c4ebb0f539eaad9868b