Fix shared lib test failures introduced by D51962231
Summary:
D51962231 introduced an issue with the shared-so tests that I missed because they ended up reporting as a warning.  The issue is that they OOM due to an infinite loop in the re-write, because we're adding and removing out of order, so we just remove the just added operand.

But there's another issue as well - we can't actually push a `MemImm` down to the auto-gen layer.  Currently when we allocate a `MemImm` we will do a `Operand.setMemoryAddress` which then sets the "operand type" to a `Mem`.  The autogen layer then treats this as a memory address and generates the wrong code.

It does appear to work if we turn these else blocks doing a transform from the memory to an immediate, but that's clearly not right as we can't disambiguate between memory and immediates anymore.  Arguably the original change is also just as broken as whether we should be producing a `Mem` or `MemImm` on the move is unclear.

This just pushes more tech debt into the T140174965.  I think the right solution here is that we need to reify `MemImm` as one of the supported `OPERAND_TYPE` that get declared, and then update auto gen appropriately.  There's probably some other fixups required because that should probably be the same as having these be initialized as immediates of `kObject` (which might be another fix, as theoretically all of the code gen for `MemImm` and `Imm` should be the same, at least on x64).

Reviewed By: jbower-fb

Differential Revision: D52515003

fbshipit-source-id: 1ebf340771231a6c7064506ef664f5b31a599a31