Fix crashing bug with large globals as first instruction in shadow code
Summary:
Pytorch is generating a function which has a large `co_names` and the first `LOAD_GLOBAL` in the function is referring to one of the names that requires an `EXTENDED_ARG`.

The normal Python compiler won't ever produce this because the first global will end up in the dictionary first, and because dictionaries are ordered, it will end up as the first entry in `co_names`.  The names that are being generated are also all going unused, which makes it extra weird from a normal Python compiled function.

This fixes an off-by-one error where we fail to go back to look at the `EXTENDED_ARG` for the first instruction.

This repro actually ends up crashing slightly earlier than the PyTorch repro in that we're crashing reading from the global cache array.  But the PyTorch repro is probably just getting luckier that wherever we're indexing happens to be valid memory, and we pull out some random value for the cache.

Reviewed By: swtaarrs

Differential Revision: D52494982

fbshipit-source-id: 9439ed57d390b896cacb40f098707199f127756c