Rerun clang-format
Summary:
It was silently broken when we moved to the CinderX/ directory.

Needed to add an import to Python.h in perf_jitdump.cpp.  It pulls in
pycore_eval.h which doesn't know about `PyObject` on its own.
clang-format is reordering the includes in perf_jitdump.cpp which
exposes this, everything was previously working because of transitive
includes.

Also removed the `__STRICTMODULE_C` hack from _strictmodule.c and
pystrictmodule.h.  pystrictmodule.h needs to unconditionally include
strict_module_checker_interface.h because it needs the declaration for
`StrictModuleChecker`. Again this is turning into an error because
clang-format is reordering the includes.

Reviewed By: brittanyrey

Differential Revision: D52541546

fbshipit-source-id: 89ae0189e999ce07adc3f0c1a42637f53bb31f75