Fix cinderjit_profile_test.py to run under jitlist_bisect.py
Summary:
Two small fixes:

First, restore `PYJIT_RESULT_NOT_ON_JITLIST`. It was added in D36730915 then
reverted in D37562166 because other parts of the diff caused problems. I used
it in D38849608 to let `failUnlessJITCompiled()` play nice with a jit-list (and
by extension, `jitlist_bisect.py`), and this diff restores that functionality.

Second, add a check to `asssertDeopts()` similar to the check added in
D38849608, allowing the test to run to completion when bisecting.

Reviewed By: alexmalyshev

Differential Revision: D52610985

fbshipit-source-id: ad03f0f1121dba263e6518b473ce77784ee438b1