Structured LIR - re-use call arg processing for args, and expand to operand formats
Summary:
Currently we have three pretty different paths for adding operands which differ in how they deal with constant values.
1) The non-structured LIR string formatting
This path will actually turn constant values into the form "1:size", and so the constant will be passed through.
2) Adding operands for calls
This path will check to see if the register has a intSpec, objectSpec, or doubleSpec, and will create a constant node.
3) Constructing an instruction with a set of operands
This will go through BlockBuilder::convertArg, which will just pass the register down (ultimately by name).  Furthermore the way this is structured it doesn't seem like it would be easy to expand the same support we have elsewhere, because we can't return an `Imm` or a `VReg` from `convertArg`, and we can't apply `constexpr` on the inputs.

Ultimately we want to get rid of code path 1, but it seems like in general we do want to preserve immediate values as we're doing the code generation, so 1 and 3 seem correct.

This modifies the `appendInstr` to go through the same path as `AppendCall` and therefore support a more straight forward code generation for simple cases (although no call sites are updated).  And then it makes `AppendCall` support all of the `OperandArg` formats via forwarding the call to `instr->addOperand`.

This still leaves instr->addOperand as being more limited then the arguments passed to `AppendCall` and `appendInstr`, but brings the 3 cases more in-line.  A potential easy follow-up would be to move all of the logic in `GenericCreateInstrInput` into `addOperand`.

As a consequence of this we now see constants showing up in places where we didn't before, and the back end code gen needs some tweaks!  In particular this happens with returns and yields.

It also happens with floating point values where we now can generate something like `FAdd {imm} {imm}`.  We'll also generate `add {imm} {imm}` but a postgen re-writer pass already fixes that up by removing one of the immediates.

For floating point values we currently don't benefit from sending down constants at all though...  We always need to generate a mov into a general purpose register, and then do a move into a floating point register.  So this just drops the support for sending down floating point values as immediates.

Ultimately this is in support of having `kIntConvert` work where the differences in code gen is currently causing compatibility issues when moving to structured LIR.

Reviewed By: alexmalyshev

Differential Revision: D52571071

fbshipit-source-id: 765d3047378766dff00b02c6b3bf9f8fdf914213