Structured LIR - Add overloads for more complicated outputs
Summary:
`appendInstr` has the pattern of accepting the destination as the first argument.  But sometimes the destination isn't just a register, and we end up passing the output as something that looks more like its in an input position.

This just allows us to pass these more complicated outputs as the first parameter before the instruction normalizing our usage of structured LIR.

Reviewed By: alexmalyshev

Differential Revision: D52573999

fbshipit-source-id: 5f49110df5197c5096f01395f48f9d4ba1f30ecd