Use structured LIR for Phi nodes
Summary:
Uses structured LIR for Phi nodes.

There's some subtle ugliness which is going on here which exists before this diff.  HIR BasicBlock nodes are casted to a LIR BasicBlock node, and then fixed up later to be the right type.  That was a little subtle before because we would just take the address of the node and pass it down.

Nothing is done to change that, but now there's a comment at least indicating that there's something funky.

Reviewed By: jbower-fb

Differential Revision: D52575308

fbshipit-source-id: 1a9b141a093cb19a42f5f251cc42e7f256a81906