Use structured LIR for incref
Summary:
Converts incref to structured LIR.  To support updating `kRefTotalAddr` we get a new appendInstr variation which supports `OutMemImm` as the first arg.

We also get a new `appendBranch` variation which supports testing flags, and only has one destination branch (with the following block being the implicit next block).  This accurately tracks our successor blocks so that we get the proper spilling done in reg alloc.

Reviewed By: jbower-fb

Differential Revision: D52577069

fbshipit-source-id: fd086aaa93913e99928fe583d1abcaa84067969f