Run JIT profile tests under plain JIT / interpreter
Summary:
First, rename the file back to `test_*` so it runs as part of the
normal test suites. This ensures the various assertions are consistent across
all execution modes.

Then, to make it pass in all modes, rename and slightly change the meanings of
the `PROFILING`/`TESTING` globals: `TESTING` was previously used to disable
calls to `assertDeopts()` when run in profiling mode, applied to broad sections
of code. I shuffled things around so the check for the new `TEST_PROFILE`
happens inside `assertDeopts()` now, letting any other assertions run in all
execution modes. There are still a few `if not MAKE_PROFILE`, to avoid
polluting the profile data with bad types, but I think this is better overall.

I also moved some types from class scope to local scope, so they're properly
reinitialized with each iteration of the test. This was exposed by the refleak
tests.

Reviewed By: jbower-fb

Differential Revision: D52712943

fbshipit-source-id: c259c1471ce5094eabc7121f8100b0ea0b2b7c06