Expose C API to enable calling module init for statically linked extensions
Summary:
Since we statically link C++ extensions into the main exe (together with the Python interpreter), we use a custom "static extension finder" to find and load such extensions from the main exe.

This relies on a static mapping from the names of statically linked extensions to their module init functions, which the custom finder calls directly during import.

Before Python 3.12, we were able to do that using several CPython private symbols that are no longer available to us in CPython 3.12. The equivalent functionality that we need in CPython 3.12 is in the `_PyImport_SwapPackageContext` function. In addition to also being a private CPython API, that function is not declared with `PyAPI_FUNC`, so it's not visible (when building with `-fvisibility=hidden`).
Since we need to patch CPython either way (e.g. to add `PyAPI_FUNC`), it seems cleaner to just add a new private C API that encapsulates the semantics we need, namely, calling a module init function via a function pointer using a provided package context.

Reviewed By: zsol

Differential Revision: D52642782

fbshipit-source-id: 7e6a97b8f1bed2bfc15da9d612279021726093d6