Remove Ci_CallDescriptorOnInvokeFunction (and fix associated issues)
Summary:
Fixes an issue where we're losing whether a function is static/class between a `_PyClassLoader_ResolveFunction` call and ` _PyClassLoader_GetIndirectPtr`.  This replaces `_PyClassLoader_GetIndirectPtr` with a version which does the resolution (which should basically be a series of dict lookups) rather than requiring the values to be passed back in.

I initially looked at simply not un-wrapping the value coming out of _PyClassLoader_ResolveFunction and then having the consumers deal with the unwrapping themselves (promoting classloader_maybe_unwrap_callable to _PyClassLoader_MaybeUnwrapCallable).  That's not too bad but it does mean both the interpreter loop and the eval loop and dealing with some finicky information they don't really care about, as they now need to get the right thing for invoking, and pass the right thing down to _PyClassLoader_GetIndirectPtr.  It also already feels a little smelly to me that we're getting some things from _PyClassLoader_ResolveFunction, and then needing to pass the combo of what we passed down to it and what we got out of it to another function, and having the unwrapper version in the picture only complicates it.

I think a cleaner API here might be that _PyClassLoader_ResolveFunction returns an object which can be further interrogated, e.g. a ResolvedFunction that then has ResolvedFunction_GetFunction, ResolvedFunction_GetIndirectPtr, etc...  It could carry with it the path, the container, etc...  But that also comes at the downside of needing to allocate to do the resolution, which is also less than ideal (although maybe handled with a singleton cache given the super short lifetime of these things).  Anyway I don't think performance is a big concern here right now so this felt like it was overall making things cleaner even if more redundant under the hood.

It then removes the `Ci_CallDescriptorOnInvokeFunction` knob now that it's enabled where static Python is in use.

Reviewed By: carljm

Differential Revision: D52525922

fbshipit-source-id: 249ac69d9568de51d2a775fbe02259824483f629